+++ pkg/web/status/getByTeam.go
@@ -18,7 +18,7 @@ func GetByTeam(ctx *gin.Context) {
 		return
 	}
 
-	entStatus, err := data.Status.GetAllByTeamWithEdges(int8(team))
+	entStatus, err := data.Status.GetAllByTeamWithEdges(team)
 	if err != nil {
 		ctx.JSON(500, gin.H{
 			"error": err.Error(),
